feat: Add resource management system for performance optimization
- Implement ResourceManager with system monitoring (CPU, memory, load average)
- Add adaptive throttling based on system load and resource limits
- Integrate resource checks into orchestrator task assignment
- Configure agent-specific limits (opencode: 3 tasks/50% CPU/2GB RAM, qwen: 2/70%/4GB, qwen-code: 2/60%/3GB)
- Add global limits (80% CPU, 85% memory, load average 4.0)
- Create resource_limits.json configuration file format
- Update orchestrator to register agents with resource manager
- Add task count tracking and throttling decisions
- Fix database inconsistencies (next_obs_id sync, cleanup old sessions)
- Update .gitignore to exclude log files
- Add sysinfo dependency for system monitoring

+++ src/core/agent_registry_ext.rs
@@ -87,7 +87,7 @@ impl AgentRegistryExt {
         )?;
 
         eprintln!("[Agent] Cleaned up {} stale sessions", modified);
-        Ok(modified as usize)
+        Ok(modified)
     }
 
     /// Get active agents count
